Reduce code reuse
